--- Code/access_manager/__init__.py
@@ -1,3 +1,0 @@
-from access_manager.access_flags import (CAN_CHANGE_ACCESS,
-                                         CAN_CHANGE_PASSWORD, CAN_DELETE_USERS)
-from access_manager.auth_manager import AuthManager
