+++ Code/api/auth/delete_user.py
@@ -1,36 +1,33 @@
-from access_manager import CAN_CHANGE_ACCESS
 from auth.bp_reg import auth_bp
 from main_impt import auth_manager
 from quart import jsonify, request
+from systems.access_manager import CAN_DELETE_USERS
 
 
-@auth_bp.route('/change_user_access', methods=['POST'])
-async def change_user_access():
+@auth_bp.route('/delete_user', methods=['POST'])
+async def api_delete_user():
     try:
         data = await request.get_json()
-
+        
+        if 'requester_token' not in data:
+            return jsonify({'message': 'Field "requester_login" is required'}), 400
+        
         if 'login' not in data:
             return jsonify({'message': 'Field "login" is required'}), 400
-
-        if 'new_access' not in data:
-            return jsonify({'message': 'Field "new_access" is required'}), 400
-
-        if 'requester_token' not in data:
-            return jsonify({'message': 'Field "requester_token" is required'}), 400
-
+        
         try:
             access = await auth_manager.get_user_access(data['requester_token'])
-            if not access & CAN_CHANGE_ACCESS:
+            if not access & CAN_DELETE_USERS:
                 return jsonify({'message': 'Access denied'}), 403
         
         except ValueError as e:
             return jsonify({'message': str(e)}), 403
-
-        await auth_manager.change_user_access(data['login'], data['new_access'])
-        return jsonify({'message': 'Access level changed successfully'}), 200
+        
+        await auth_manager.delete_user(data['requester_login'])
+        return jsonify({'message': 'User deleted successfully'}), 200
 
     except ValueError as e:
         return jsonify({'message': str(e)}), 403
-
+    
     except Exception as e:
         return jsonify({'message': 'An unexpected error occurred', 'error': str(e)}), 500
