+++ Code/api/auth/logout_user.py
