--- Code/auth/delete_user.py
@@ -1,33 +1,0 @@
-from access_manager import CAN_DELETE_USERS
-from auth.bp_reg import auth_bp
-from main_impt import auth_manager
-from quart import jsonify, request
-
-
-@auth_bp.route('/delete_user', methods=['POST'])
-async def api_delete_user():
-    try:
-        data = await request.get_json()
-        
-        if 'requester_token' not in data:
-            return jsonify({'message': 'Field "requester_login" is required'}), 400
-        
-        if 'login' not in data:
-            return jsonify({'message': 'Field "login" is required'}), 400
-        
-        try:
-            access = await auth_manager.get_user_access(data['requester_token'])
-            if not access & CAN_DELETE_USERS:
-                return jsonify({'message': 'Access denied'}), 403
-        
-        except ValueError as e:
-            return jsonify({'message': str(e)}), 403
-        
-        await auth_manager.delete_user(data['requester_login'])
-        return jsonify({'message': 'User deleted successfully'}), 200
-
-    except ValueError as e:
-        return jsonify({'message': str(e)}), 403
-    
-    except Exception as e:
-        return jsonify({'message': 'An unexpected error occurred', 'error': str(e)}), 500
