--- Code/db_manager/__init__.py
@@ -1,4 +1,0 @@
-from db_manager.async_DB import (AsyncDB, CheckConstraintError,
-                                 ForeignKeyConstraintError,
-                                 NotNullConstraintError, UniqueConstraintError)
-from db_manager.settings_manager import SettingsManager
