--- Code/factory/__init__.py
@@ -1,4 +1,0 @@
-from factory.base_component import Component
-from factory.base_entity import Entity
-from factory.errors import PrototypeError
-from factory.prototype_factory import PrototypeFactory
