+++ Code/main.py
@@ -6,11 +6,11 @@ import platform
 import subprocess
 import sys
 
-from auth import auth_bp
-from auto_updater import needs_update
+from api.auth import auth_bp
 from colorlog import ColoredFormatter
-from db_manager import SettingsManager
 from quart import Quart
+from systems.auto_updater import needs_update
+from systems.db_manager import SettingsManager
 
 app = Quart(__name__)
