+++ Code/main_impt.py
@@ -1,5 +1,5 @@
-from access_manager import AuthManager
-from factory import PrototypeFactory
+from systems.access_manager import AuthManager
+from systems.entiti_factory import PrototypeFactory
 
 auth_manager = AuthManager()
 prototype_factory = PrototypeFactory()
