+++ Code/systems/access_manager/__init__.py
@@ -1,0 +1,4 @@
+from systems.access_manager.access_flags import (CAN_CHANGE_ACCESS,
+                                                 CAN_CHANGE_PASSWORD,
+                                                 CAN_DELETE_USERS)
+from systems.access_manager.auth_manager import AuthManager
