+++ Code/systems/auto_updater/__init__.py
@@ -1,0 +1,1 @@
+from systems.auto_updater.auto_updater import needs_update
