+++ Code/systems/entiti_factory/base_component.py
