+++ Code/systems/entiti_factory/base_entity.py
