+++ Code/systems/entiti_factory/test_fold/component.py
@@ -1,4 +1,4 @@
-from factory.base_component import Component
+from systems.entiti_factory.base_component import Component
 
 
 class HealthComponent(Component):
